Move simplifyDistricts to merge in worker
