+++ src/manage/src/commands/create-random-projects.ts
@@ -7,7 +7,6 @@ import _ from "lodash";
 import { connectionOptions } from "../lib/dbUtils";
 import { RegionConfig } from "../../../server/src/region-configs/entities/region-config.entity";
 import { Project } from "../../../server/src/projects/entities/project.entity";
-import { simplifyDistricts } from "../../../server/src/projects/controllers/projects.controller";
 import { TopologyService } from "../../../server/src/districts/services/topology.service";
 import { WorkerPoolService } from "../../../server/src/districts/services/worker-pool.service";
 import { User } from "../../../server/src/users/entities/user.entity";
@@ -38,8 +37,6 @@ export default class CreateRandomProjects extends Command {
     const regionConfigRepo = connection.getRepository(RegionConfig);
     const projectRepo = connection.getRepository(Project);
     const userRepo = connection.getRepository(User);
-
-    const boundSimplifyDistricts = simplifyDistricts.bind(this);
 
     const regions = await regionConfigRepo.find(
       args.region === "all"
@@ -85,12 +82,14 @@ export default class CreateRandomProjects extends Command {
       }
       const lockedDistricts = new Array(numberOfDistricts).fill(false);
       const numberOfMembers = new Array(numberOfDistricts).fill(1);
-      const districts = await geoCollection.merge({
-        districtsDefinition,
-        numberOfDistricts,
-        user,
-        regionConfig: region
-      });
+      const { districts, simplifiedDistricts } = {
+        ...(await geoCollection.merge({
+          districtsDefinition,
+          numberOfDistricts,
+          user,
+          regionConfig: region
+        }))
+      };
       if (!districts) {
         this.log(`Could not generate geojson`);
         this.exit(1);
@@ -100,7 +99,7 @@ export default class CreateRandomProjects extends Command {
       project.regionConfig = region;
       project.districtsDefinition = districtsDefinition;
       project.districts = districts;
-      project.simplifiedDistricts = boundSimplifyDistricts(districts);
+      project.simplifiedDistricts = simplifiedDistricts;
       project.lockedDistricts = lockedDistricts;
       project.numberOfMembers = numberOfMembers;
       project.populationDeviation = 5;
